refactor(auth): extract verifyProjectAuth, drop the project direct seam
Reuse the project route factory's seam for GET /api/public/projects and delete verifyProjectAuthDirect; the org direct seam is now the only one in shadow.direct.
